smart_planner_web_backend/requirements.txt
# FastAPI + server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic + settings
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Auth / security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Firebase
firebase-admin==6.2.0
python-dotenv==1.0.1

# Rate limiting
slowapi==0.1.9

# AI providers
openai==1.3.2
httpx==0.27.2
google-generativeai==0.3.0
tenacity==8.2.3
